BACK/my_docker_django_app/requirements.txt
asgiref==3.8.1
#audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
idna==3.10
mysqlclient==2.2.7
numpy==2.2.3
pydub==0.25.1
python-core==0.0.1
requests==2.32.3
setuptools==75.8.0
SpeechRecognition==3.14.1
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-cors-headers==3.7.0
djangorestframework
djangorestframework-simplejwt
